requirements.txt
XInput-Python
vgamepad
strenum